Disable TLS by default
OTEL_EXPORTER_OTLP_INSECURE is an environment variable understood by most gRPC trace exporter clients. In most cases, the sidecar injected don't setup a TLS certificate as the connection is local to the pod. A user can however, set this to false again, if they use TLS in their setup for the local tracing sidecar.